♻️ Refactor: TabBar ui, 훅 로직 분리
